Refactored the last components. Fixed the bug regarding referral view. Fixed error message from console.
