+++ ui/src/components/modals/FormModal.tsx
@@ -1,24 +1,13 @@
 import React, { SetStateAction } from "react";
-import { Choice, ContractId, Time } from "@daml/types";
+import { Choice, ContractId } from "@daml/types";
 import { Event as DEvent, CreateEvent } from "@daml/ledger";
 import { useLedger } from "@daml/react";
-
-import FormikMod, {
-  Formik,
-  Form,
-  Field,
-  FormikHelpers,
-  FieldAttributes,
-  useField,
-} from "formik";
+import FormikMod, { Formik, Form, FormikHelpers } from "formik";
 import { ArrowRight, Check, X } from "phosphor-react";
 import { Link } from "react-router-dom";
-import Select from "react-select";
-import Modal from "./Modal";
-import DayPicker from "./DayPicker";
-import TimePicker from "react-time-picker";
+import Modal from "./ModalLayout";
+import { Nothing } from "utils";
 
-export const Nothing = Symbol("Nothing");
 // eslint-disable-next-line @typescript-eslint/no-redeclare
 type Nothing = typeof Nothing;
 
@@ -56,7 +45,7 @@ interface Failure<C> {
   error: any;
 }
 
-type ChoiceModalProps<T extends object, C, R, K> = {
+type FormModalProps<T extends object, C, R, K> = {
   choice: Choice<T, C, R, K>;
   contract: ContractId<T>;
   submitTitle: string;
@@ -91,7 +80,7 @@ export const SubmitButton: React.FC<{
   </button>
 );
 
-export function ChoiceModal<T extends object, C, R, K>({
+export function FormModal<T extends object, C, R, K>({
   choice,
   contract,
   submitTitle,
@@ -102,7 +91,7 @@ export function ChoiceModal<T extends object, C, R, K>({
   successWidget,
   failureWidget,
   children,
-}: ChoiceModalProps<T, C, R, K>) {
+}: FormModalProps<T, C, R, K>) {
   const [modalActive, setModalActiveInner] = React.useState(false);
   const [successOrFailure, setSuccessOrFailure] =
     React.useState<MaybeSuccessOrFailure<C, R>>(Nothing);
@@ -120,8 +109,6 @@ export function ChoiceModal<T extends object, C, R, K>({
     values: PartialMaybe<C>,
     { setSubmitting }: FormikHelpers<PartialMaybe<C>>
   ) => {
-    console.log("Exercising option: " + choice);
-    console.log(values);
     const arg = complete(values);
     if (arg) {
       const success = (a: [R, DEvent<object>[]]) => {
@@ -179,7 +166,12 @@ export function ChoiceModal<T extends object, C, R, K>({
     }
   } else {
     content = (
-      <Formik initialValues={initialValues} onSubmit={submitF}>
+      <Formik
+        initialValues={initialValues}
+        onSubmit={submitF}
+        validateOnChange={false}
+        validateOnBlur={false}
+      >
         {({ errors, touched, isSubmitting }) => (
           <Form className={className}>
             {typeof children === "function"
@@ -215,167 +207,6 @@ export function ChoiceModal<T extends object, C, R, K>({
     </>
   );
 }
-
-export type ChoiceErrorsType = { [_: string]: string | undefined };
-
-export const validateNonEmpty = (label: string) => (a: any) => {
-  let error;
-  if (a === Nothing) {
-    error = `${label} is required`;
-  }
-  return error;
-};
-
-export const RenderError: React.FC<{ error: string | undefined }> = ({
-  error,
-}) => <>{error && <div className="text-sm text-red-800">{error}</div>}</>;
-
-export const LField: React.FC<
-  FieldAttributes<any> & {
-    label: string;
-    errors?: ChoiceErrorsType;
-  }
-> = ({ errors, label, ...props }) => {
-  const error = errors?.[props.name];
-  return (
-    <div className="">
-      <label htmlFor={props.name} className="block label-sm ">
-        {label}
-      </label>
-      <Field
-        {...props}
-        className="bg-trueGray-100 h-11 rounded w-full"
-        validate={validateNonEmpty(label)}
-      />
-      <RenderError error={error} />
-    </div>
-  );
-};
-
-export const EField: React.FC<{
-  name: string;
-  e: any;
-  label: string;
-  errors?: ChoiceErrorsType;
-}> = ({ name, e, label, errors }) => {
-  const [, , helpers] = useField({
-    name,
-    validate: validateNonEmpty(label),
-  });
-  const { setValue } = helpers;
-  const error = errors?.[name];
-  return (
-    <div className="flow flow-col">
-      <label htmlFor={name} className="block label-sm">
-        {label}
-      </label>
-      <Select
-        cla1ssNamePrefix="react-select-modal-enum"
-        multi={false}
-        options={e.keys.map((a: string) => ({ value: a, label: a }))}
-        onChange={(option) => setValue(option?.value)}
-        styles={{
-          singleValue: () => ({ textOverflow: "ellipsis", maxWidth: "10em" }),
-        }}
-        validate={undefined}
-      />
-      <RenderError error={error} />
-    </div>
-  );
-};
-
-export const DayPickerField: React.FC<{
-  name: string;
-  errors?: ChoiceErrorsType;
-}> = ({ name, errors }) => {
-  const [field, , { setValue }] = useField<string | Nothing>(name);
-  const error = errors?.[name];
-  return (
-    <>
-      <DayPicker
-        date={
-          field.value === Nothing
-            ? new Date()
-            : new Date(field.value + "T00:10:00")
-        }
-        setDate={(d) => setValue(d.toISOString().split("T")[0])}
-        theme={{ blue: "var(--blue)" }}
-      />
-      <RenderError error={error} />
-    </>
-  );
-};
-
-// convert a JavaScript "Date" value into a DAML-backend "Time" value
-const dateToTime = (d: Date): Time => d.toISOString();
-
-export const DayTimePickerField: React.FC<{
-  name: string;
-  errors?: ChoiceErrorsType;
-}> = ({ name, errors }) => {
-  const [field, , { setValue }] = useField<Time | Nothing>(name);
-  if (field.value === Nothing) {
-    // can't help but render some picked date, so might as well set it
-    const f = new Date();
-    console.log("default combined", f);
-    setValue(dateToTime(f));
-  }
-  const defaultField =
-    field.value === Nothing ? new Date() : new Date(field.value);
-  let date = new Date(
-    defaultField.getFullYear(),
-    defaultField.getMonth(),
-    defaultField.getDate()
-  );
-  let time = new Date(
-    0,
-    0,
-    0,
-    defaultField.getHours(),
-    defaultField.getMinutes(),
-    defaultField.getSeconds()
-  );
-  const updateField = () => {
-    const f = new Date(
-      date.getFullYear(),
-      date.getMonth(),
-      date.getDate(),
-      time.getHours(),
-      time.getMinutes(),
-      time.getSeconds()
-    );
-    console.log("new combined", f);
-    setValue(dateToTime(f));
-  };
-  const error = errors?.[name];
-  return (
-    <>
-      <DayPicker
-        date={date}
-        setDate={(d) => {
-          console.log("new date", d);
-          date = d;
-          updateField();
-        }}
-        theme={{ blue: "var(--blue)" }}
-      />
-      <TimePicker
-        onChange={(t) => {
-          console.log("new time", t);
-          if (t instanceof Date) {
-            time = t;
-          } else {
-            const [hours, minutes] = t.split(":");
-            time = new Date(0, 0, 0, parseInt(hours), parseInt(minutes));
-          }
-          updateField();
-        }}
-        value={time}
-      />
-      <RenderError error={error} />
-    </>
-  );
-};
 
 export const creations: (_: DEvent<object>[]) => CreateEvent<object>[] = (
   evts
