+++ ui/src/utils/index.ts
@@ -1,7 +1,7 @@
 import { useState, useMemo } from "react";
 import dateFormat from "dateformat";
 
-function intercalate<X>(xs: X[], sep: X) {
+export function intercalate<X>(xs: X[], sep: X) {
   return xs.flatMap((x) => [sep, x]).slice(1);
 }
 
@@ -14,14 +14,17 @@ export function* mapIter<A, B>(
   }
 }
 
-function leftJoin<K, X, Y>(
+export function leftJoin<K, X, Y>(
   xs: Map<K, X>,
   ys: Map<K, Y>
 ): Map<K, [X, Y | undefined]> {
   return new Map(mapIter(([k, x]) => [k, [x, ys.get(k)]], xs.entries()));
 }
 
-function innerJoin<K, X, Y>(xs: Map<K, X>, ys: Map<K, Y>): Map<K, [X, Y]> {
+export function innerJoin<K, X, Y>(
+  xs: Map<K, X>,
+  ys: Map<K, Y>
+): Map<K, [X, Y]> {
   let ret = new Map();
   for (const [k, x] of xs.entries()) {
     const y = ys.get(k);
@@ -35,7 +38,7 @@ export const formatDate = (d: Date) => dateFormat(d, "ddd, mmm d, yyyy");
 // Requirement: do not pass an array as "memoKeys" argument,
 //  since it's assumed that "useMemo" doesn't look at the content
 //  of arrays.
-function useAsync<T>(f: () => Promise<T>, memoKeys: any): T | null {
+export function useAsync<T>(f: () => Promise<T>, memoKeys: any): T | null {
   const [[v, lastMemoKeys], setV] = useState<[T | null, any]>([null, null]);
   useMemo(
     // some false positives so we do the extra comparison to avoid extra `setV`
@@ -49,4 +52,12 @@ function useAsync<T>(f: () => Promise<T>, memoKeys: any): T | null {
   return v;
 }
 
-export { innerJoin, leftJoin, intercalate, useAsync };
+export const Nothing = Symbol("Nothing");
+
+export const validateNonEmpty = (label: string) => (a: any) => {
+  let error;
+  if (a === Nothing) {
+    error = `${label} is required`;
+  }
+  return error;
+};
