+++ ui/src/views/bills/Bills.tsx
@@ -1,16 +1,21 @@
 import React from "react";
-import { TabularView } from "components/TabularScreen";
+import TabularView from "components/TabularScreen";
 import { useBills } from "hooks/bills";
 
+/**
+ * function that is passed in the TableBills component
+ * which retrieves bills data through the useBills hook.
+ * See "hooks/bills" for more details
+ */
 const useBillsData = () => useBills({});
 
+/** Table view for bills */
 const TableBills: React.FC = () => {
   return (
     <TabularView
       title="Bills"
       useData={useBillsData}
       fields={[
-        // NB: outputs provider party name (e.g. "Radiologist") instead of human-friendly provider name (e.g. "Beta Imaging Labs").
         { label: "Provider", getter: (o) => o.bill?.payload?.provider },
         { label: "Amount", getter: (o) => o?.bill?.payload?.amount },
         {
@@ -19,8 +24,8 @@ const TableBills: React.FC = () => {
         },
         { label: "Paid", getter: (o) => (o?.receipt?.payload ? "Yes" : "No") },
       ]}
-      tableKey={(o) => o.bill.contractId}
-      itemUrl={(o) => o.bill.contractId}
+      tableKey={(o) => o?.bill?.contractId}
+      itemUrl={(o) => o?.bill?.contractId}
     />
   );
 };
